CON-1395: Avoid FailedPreStopHook warning on hpe-csi-node termination
The csi-node-driver-registrar container in the hpe-csi-node daemonset defines
a pre-stop hook that runs a command in the container.  However, the container
does not include a shell command, so the hook execution fails with this error:
"exec: \"/bin/sh\": stat /bin/sh: no such file or directory".  Avoid the error
by removing the hook.

Signed-off-by: Roger Kumpf <roger.kumpf@hpe.com>