Avoid FailedPreStopHook warning on hpe-csi-node termination (#194)
The csi-node-driver-registrar container in the hpe-csi-node daemonset defines
a pre-stop hook that runs a command in the container.  However, the container
does not include a shell command, so the hook execution fails with this error:
"exec: \"/bin/sh\": stat /bin/sh: no such file or directory".  Avoid the error
by removing the hook.

Update to the csi-node-driver-registrar 2.0.1 release, which cleans up the
domain socket that was meant to be removed by the pre-stop hook.